pyproject.toml
[project]
name = "mavka"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
